requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
nest-asyncio
pyngrok
python-dotenv
pandas
numpy
scikit-learn
prophet